🐳 Make backend in boxer.sh script configurable
The changes made to the boxer.sh script allow for the backend to be configured. Previously, Docker was hardcoded as the backend for running containers. Now, a variable `BACKEND` has been introduced which defaults to "docker", but can be changed to other container runtimes like "podman" or "nerdctl". This change provides more flexibility and allows users to choose their preferred container runtime.